Solution for the Lowest Common Ancestor Problem
I thought this might be an interesting solution to this problem. In my experience, it can be treated in two ways: as if a query for a Lowest Common Ancestor is done once or as if it will be repeated on the same tree. I imagine that, in practice, if such a function were needed, it would do more than one query on the tree. When I first studied the LCA problem, this is one of the best solutions I found for repeated queries, due to the fact that, after a precalculation done in O(n log(n)) time (n being the number of nodes), every query will be done in O(1) time. A detailed explanation of this solution can be found here:
https://www.topcoder.com/community/data-science/data-science-tutorials/range-minimum-query-and-lowest-common-ancestor/
The solution could have been improved even further by using Segment Trees for RMQ. That way, the tree could also be modified and the solution for LCA would be dynamic: the segment tree will be modified in O(log n) according to the changes in the original tree and the LCA queries will still return correct results in O(1). However, the assignment did not specify that the tree would be modified, so I did not implement this; I just thought it would be interesting to mention.