Solution.cpp
I did not understand how to use unique_ptr if some functions return a pointer(for example 47: Next() ) and then the rule of individual ownership will be violated. I decided to use a shared_ptr. Or does it make no sense because there will still be a memory leak? I'm not sure.